project final commit

+++ blogs/app/Http/Controllers/PostsController.php
@@ -1,12 +1,21 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Support\Facades\storage;
 use Illuminate\Http\Request;
 use App\Models\Post;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -39,13 +48,31 @@ class PostsController extends Controller
         $request->validate([
             'title' => 'required',
             'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+        // handle File Upload
+        if ($request->hasFile('cover_image')) {
+            // Get filename with extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathInfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
         // create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
-        return redirect('/posts')->with('success', 'Post Created');
+        return redirect('/home')->with('success', 'Post Created');
         
         
     }
@@ -71,6 +98,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
         return view('posts.edit')->with('post', $post);
 
     }
@@ -88,12 +121,29 @@ class PostsController extends Controller
             'title' => 'required',
             'body' => 'required',
         ]);
+
+        if ($request->hasFile('cover_image')) {
+            // Get filename with extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathInfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+
         // create post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
-        return redirect('/posts')->with('success', 'Post Updated');
+        return redirect('/home')->with('success', 'Post Updated');
     }
 
     /**
@@ -105,7 +155,16 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = POST::find($id);
+        // check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
+        if ($post->cover_image != 'noimage.jpg') {
+            // Delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
-        return redirect('/posts')->with('success', 'Post Deleted');
+        return redirect('/home')->with('success', 'Post Deleted');
     }
 }
